relaxing flere vagter

+++ Dyreinternattet Semesterprojekt Vinter 2023/Models/Vagtplan/Vagt.cs
@@ -5,14 +5,14 @@ namespace Dyreinternattet_Semesterprojekt_Vinter_2023.Models.Vagtplan
     public class Vagt
     {
 		public Medarbejder Worker { get; set; }
-		public int StartTid { get; set; }
-		public int SlutTid { get; set; }
+		public double StartTid { get; set; }
+		public double SlutTid { get; set; }
 
 		public Vagt()
 		{
 		}
 
-		public Vagt(Medarbejder worker, int starttid, int sluttid)
+		public Vagt(Medarbejder worker, double starttid, double sluttid)
 		{
 			Worker = worker;
 			StartTid = starttid;
